+++ CHANGELOG.md
@@ -5,7 +5,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
+## Fixed
+* Correctly display activity duration as timedelta formatted as `HH:MM:SS` instead of
+  raw seconds in time series plots on activity page.
+## Changed
+* #189: Upgrade `fitparse` to most recent version and use `enhanced_altitude` instead of
+  `altitude` value when parsing fit files, since `enhanced_altitude` seems to contain
+  the correct altitude value for all garmin devices. This requires a reimport of
+  activities.
+## Added
+* Use value of `altitude` field in case `enhanced_altitude` field is not available during
+  parsing of fit files.
 * Test both mounting Block and MTP devices.
 
 ## [0.21.1](https://github.com/fgebhart/workoutizer/releases/tag/v0.21.1) - 2021-07-17